+++ docs_website/docusaurus.config.js
@@ -40,10 +40,7 @@ const config = {
             /** @type {import('@docusaurus/preset-classic').Options} */
             ({
                 docs: {
-                    sidebarPath: require.resolve("./sidebars.js"),
-                    // Please change this to your repo.
-                    // Remove this to remove the "edit this page" links.
-                    editUrl: "https://github.com/CARTAvis/carta-frontend"
+                    sidebarPath: require.resolve("./sidebars.js")
                 },
                 theme: {
                     customCss: require.resolve("./src/css/custom.css")