+++ CHANGELOG.md
@@ -9,16 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added a shortcut button for image annotation ([#2167](https://github.com/CARTAvis/carta-frontend/issues/2167)).
 * Added support for AIPS beam images ([#2164](https://github.com/CARTAvis/carta-frontend/issues/2164)).
+* Added telemetry data for spectral profile, PV, moment, and catalog creation. ([#2208](https://github.com/CARTAvis/carta-frontend/issues/2208)).
 * Added an increase to the default compression ratio to 32 when the header unit of the image is km/s ([#1223](https://github.com/CARTAvis/carta-frontend/issues/1223)).
 * Added a toggle in the moment generator to match generated image(s) ([#2121](https://github.com/CARTAvis/carta-frontend/issues/2121)).
 * Added automatically generated documentation website.
+* Added the ability to share a workspace via URL when using the frontend via the controller.
 ### Changed
 * Changed the default title string in the image viewer ([#2168](https://github.com/CARTAvis/carta-frontend/issues/2168)).
 * Modified text annotation textbox to stay the same dimension as user zoom the image ([#2162](https://github.com/CARTAvis/carta-frontend/issues/2162)).
 * Disabled spell check in text input ([#2138](https://github.com/CARTAvis/carta-frontend/issues/2138)).
 * Support animation playback with matched images in multi-panel view ([#1860](https://github.com/CARTAvis/carta-frontend/issues/1860)).
-* Modified spectral matching type selector in Image List Settings won't affact global preferences ([#2098](https://github.com/CARTAvis/carta-frontend/issues/2098)).
+* Modified spectral matching type selector in Image List Settings won't affect global preferences ([#2098](https://github.com/CARTAvis/carta-frontend/issues/2098)).
 * Enabled enter key to trigger filtering in the filter fields, and modified the buttons in the spectral line query widget and catalog widget ([#2046](https://github.com/CARTAvis/carta-frontend/issues/2046)).
+* Switched the Splatalogue URL from test to production ([#2238](https://github.com/CARTAvis/carta-frontend/issues/2238)).
+* Changed image fitting center position error unit from arcsecond to second when the format is H:M:S; added integrated flux information in the fitting result.
 ### Fixed
 * Fixed bug where line region computation width cannot be changed in spatial profile setting widget ([#2000](https://github.com/CARTAvis/carta-frontend/issues/2000)).
 * Fixed when multiple images are open, PV generator can only produce PV preview with live update for one of the images ([#2171](https://github.com/CARTAvis/carta-frontend/issues/2171)).